routes/requirements.txt
Flask==2.2.5
Flask-Cors==3.0.10
Flask-JWT==0.2.0
Flask-JWT-Extended==4.6.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
gunicorn==20.1.0
marshmallow==3.18.0
marshmallow-sqlalchemy==0.25.0
SQLAlchemy==1.4.15
psycopg2-binary
werkzeug==2.2.2
pytest
pytest-cov
requests
google-cloud-pubsub